% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Compute the euclidean algorithm for two numbers}
\usage{
euclidean(a, b)
}
\arguments{
\item{c(a, b)}{Numeric scalar or integer.}
}
\value{
The greatest common divisor of \code{a} and \code{b}.
}
\description{
Compute the euclidean algorithm for two numbers
}
\examples{
 
euclidean(123612, 13892347912) 
euclidean(100, 1000) 
}
