\name{Lab1HugoOtto-package}
\alias{Lab1HugoOtto-package}
\alias{Lab1HugoOtto}
\docType{package}
\title{
\packageTitle{Lab1HugoOtto}
}
\description{
\packageDescription{Lab1HugoOtto}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab1HugoOtto}
\packageIndices{Lab1HugoOtto}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Lab1HugoOtto}

Maintainer: \packageMaintainer{Lab1HugoOtto}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
