% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab1HugoOtto.R
\docType{package}
\name{Lab1HugoOtto}
\alias{Lab1HugoOtto}
\alias{Lab1HugoOtto-package}
\title{Lab1HugoOtto: A package containing the functions euclidean and dijkstra}
\description{
This package contains two functions:
euclidean and dijkstra.
}
\section{Euclidean}{

The euclidean function takes two numbers and returns the greatest common divisor using the euclidean algorithm.
}

\section{Dijkstra}{

The dijkstra function takes a graph and an inital node and returns the shortest path from the inital node to every other node in the graph using Dijkstra's algorithm.
}

