% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Compute the euclidean algorithm for two numbers}
\usage{
euclidean(a, b)
}
\arguments{
\item{(a, b)}{Numeric scalar or integer.}
}
\value{
The greatest common divisor of \code{a} and \code{b}.
}
\description{
Compute the euclidean algorithm for two numbers
}
\details{
This function computes euclidean algorithm to find the greatest common divisor between two integers (\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Wikipedia}).
}
\examples{
 
euclidean(123612, 13892347912) 
euclidean(100, 1000) 
}
